removes all well-trained constant effects
Except for the moodlet that is, but that's not as annoying.

Blushing and jittering are now part of the snap effects instead of being constantly applied when a dom is near. In addition, it'll tell you only ONCE when a dom is near you, and then not do it again until you've been enough time away from a dom