edited the whole fucking base
for this whole fucking thing to work correctly I had to increase the spawn rate for /obj/effect/mob_spawn/human/lavaland_syndicate/comms/space to 100% because otherwise it would have a 1% of the base to spawn with two spaces for ghosts, and 19% to spawn with any space for ghosts.

I also had to edit the non-modular map because it also is able to spawn in space and it would be the old one. I have also added listeningstation.dmm.old in non-modular, and syndielistenspace.dmm.old in modular to have the original one still in the files so that it can be easily rolled back to, if the need arises.